requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
httpx>=0.24.0
python-dotenv>=0.19.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
newsapi-python>=0.2.7
feedparser>=6.0.10
python-dateutil>=2.8.2
gunicorn>=20.1.0
